chore: nit fixes + refactor
