+++ utils/index.ts
@@ -1,5 +1,5 @@
 export * from './constants';
-export * from './css';
+export * from './dimensions';
 export * from './pkce';
 export * from './shaders';
 export * from './spotify';
